# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Proxy Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/access/check": {
            "get": {
                "tags": [
                    "access"
                ],
                "summary": "Check IP access",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP address to evaluate. Defaults to the caller IP when omitted",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Captcha token for greylist verification",
                        "name": "captcha",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.AccessDecision"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.AccessDecision": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "matchedValue": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "requiresCaptcha": {
                    "type": "boolean"
                },
                "ruleId": {
                    "type": "string"
                },
                "ruleType": {
                    "$ref": "#/definitions/core.AccessRuleType"
                }
            }
        },
        "core.AccessRuleType": {
            "type": "string",
            "enum": [
                "allow",
                "deny",
                "grey"
            ],
            "x-enum-varnames": [
                "AccessRuleAllow",
                "AccessRuleDeny",
                "AccessRuleGrey"
            ]
        },
...
    }
}